requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
datasets==2.5.2
debugpy==1.6.2
decorator==5.1.1
dill==0.3.5.1
einops==0.4.1
entrypoints==0.4
evaluate==0.2.2
executing==0.9.1
fastjsonschema==2.16.2
filelock==3.8.0
flatbuffers==1.12
fonttools==4.34.4
frozenlist==1.3.0
fsspec==2022.7.1
gast==0.4.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.0
h5py==3.7.0
huggingface-hub==0.10.0
idna==3.3
imageio==2.20.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.15.1
ipython==8.4.0
ipywidgets==8.0.2
jedi==0.18.1
joblib==1.1.0
jsonschema==4.16.0
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-widgets==3.0.3
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.1
llvmlite==0.39.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
multidict==6.0.2
multiprocess==0.70.13
nbformat==5.6.1
nd2reader==3.3.0
nest-asyncio==1.5.5
numba==0.56.0
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
PIMS==0.6.1
pims-nd2==1.1
pkgutil_resolve_name==1.3.10
plotly==5.10.0
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
Pygments==2.12.0
pynndescent==0.5.7
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytorch-lightning==1.6.5
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
scikit-learn==1.1.1
scipy==1.9.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
slicerator==1.1.0
stack-data==0.3.0
tenacity==8.1.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.1
torch-cluster==1.6.0
torch-tb-profiler==0.4.0
torchaudio==0.12.1
torchmetrics==0.9.3
torchsampler==0.1.2
torchsummary==1.5.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.22.2
typing_extensions==4.3.0
umap-learn==0.5.3
urllib3==1.26.11
vit-pytorch==0.35.8
wcwidth==0.2.5
Werkzeug==2.2.1
widgetsnbextension==4.0.3
wrapt==1.14.1
xmltodict==0.13.0
xxhash==3.0.0
yarl==1.7.2
zipp==3.8.1
